pyproject.toml
[project]
name = "biomed-exp-agent"
version = "1.0.0"
description = "BioMed-Exp Agent 4.0-R - 生物医学实验智能体系统"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "BioMed Team"}
]
keywords = ["biomedical", "agent", "experiment", "LLM", "computer-vision"]

dependencies = [
    # Agent Framework
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # LLM Providers
    "zhipuai>=2.0.0",           # GLM (智谱 AI)
    "langchain-anthropic>=0.3.0", # Claude (可选)
    "openai>=1.0.0",            # OpenAI / Ollama 兼容

    # LLM Router
    "litellm>=1.0.0",

    # Memory
    "mem0ai>=1.0.0",

    # MCP Protocol
    "mcp>=1.0.0",

    # API Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "chromadb>=0.5.0",

    # Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.0",

    # CV & ML
    "torch>=2.0.0",
    "torchvision>=0.17.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "scipy>=1.12.0",
    "scikit-image>=0.22.0",

    # Task Queue
    "celery>=5.3.0",

    # Storage
    "minio>=7.2.0",

    # UI
    "gradio>=4.0.0",

    # Utilities
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ===========================================
# uv 配置
# ===========================================
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[tool.pytest.ini_options]
